extract content to template expression
