+++ packages/adapter-utils/index.ts
@@ -24,3 +24,4 @@ export * from './src/important_values'
 export * from './src/additional_properties'
 export * from './src/config_options'
 export * from './src/error_messages'
+export * from './src/html_parser'
