+++ packages/adapter-utils/package.json
@@ -34,7 +34,9 @@
     "@salto-io/adapter-api": "0.5.0",
     "@salto-io/logging": "0.5.0",
     "@salto-io/lowerdash": "0.5.0",
+    "domhandler": "^4.2.2",
     "fast-safe-stringify": "^2.0.7",
+    "htmlparser2": "^7.2.0",
     "joi": "^17.4.0",
     "lodash": "^4.17.21",
     "object-hash": "^3.0.0",
