+++ packages/zendesk-adapter/package.json
@@ -40,9 +40,7 @@
     "@salto-io/parser": "0.5.0",
     "acorn": "^8.11.3",
     "acorn-walk": "^8.3.2",
-    "domhandler": "^4.2.2",
     "form-data": "^4.0.0",
-    "htmlparser2": "^7.2.0",
     "joi": "^17.4.0",
     "jszip": "^3.10.1",
     "lodash": "^4.17.21",
