+++ packages/zendesk-adapter/test/filters/template_engines/html_parser.test.ts
@@ -9,72 +9,8 @@ import { ElemID, InstanceElement, ObjectType, ReferenceExpression } from '@salto
 import { ARTICLE_TYPE_NAME, BRAND_TYPE_NAME, ZENDESK } from '../../../src/constants'
 import {
   parseUrlPotentialReferencesFromString,
-  parseTagsFromHtml,
   parseHtmlPotentialReferences,
 } from '../../../src/filters/template_engines/html_parser'
-
-describe('parseTagsFromHtml', () => {
-  describe('should extract URLs from HTML tags', () => {
-    it('should extract URLs from HTML content', () => {
-      const htmlContent = `
-      <a href="https://example.com">Link 1</a>
-      <img src="https://example.com/image.jpg" alt="Image">
-      <link href="https://example.com/styles.css" rel="stylesheet">
-    `
-      const { urls } = parseTagsFromHtml(htmlContent)
-      expect(urls).toEqual([
-        {
-          value: 'https://example.com',
-          loc: {
-            end: 35,
-            start: 16,
-          },
-        },
-        {
-          value: 'https://example.com/image.jpg',
-          loc: {
-            end: 93,
-            start: 64,
-          },
-        },
-        {
-          value: 'https://example.com/styles.css',
-          loc: {
-            end: 156,
-            start: 126,
-          },
-        },
-      ])
-    })
-
-    it('should handle empty HTML content', () => {
-      const htmlContent = ''
-      const { urls } = parseTagsFromHtml(htmlContent)
-      expect(urls).toEqual([])
-    })
-
-    it('should handle HTML content without any URLs', () => {
-      const htmlContent = `
-      <div>Hello, world!</div>
-      <p>This is a paragraph.</p>
-    `
-      const { urls } = parseTagsFromHtml(htmlContent)
-      expect(urls).toEqual([])
-    })
-  })
-
-  describe('should extract text from script tags', () => {
-    it('should extract text from script tags', () => {
-      const htmlContent = `
-      <script>
-        const someVar = 'some value';
-      </script>
-    `
-      const { scripts } = parseTagsFromHtml(htmlContent)
-      expect(scripts.map(s => s.value.trim())).toEqual(["const someVar = 'some value';"])
-    })
-  })
-})
 
 describe('parseUrlPotentialReferencesFromString', () => {
   let urlBrandInstance: InstanceElement
